requirements.txt
fastapi
uvicorn
tavily-python
    